+++ package.json
@@ -12,6 +12,7 @@
     "chart.js": "^4.4.2",
     "firebase": "^10.10.0",
     "firebaseui": "^6.1.0",
+    "loadash": "^1.0.0",
     "lodash": "^4.17.21",
     "vue": "^3.4.15",
     "vue-chartjs": "^5.3.0",
